requirements.txt
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
openpyxl==3.1.5
minio==7.2.8
pyyaml==6.0.2
python-dotenv==1.0.1
pytest==8.3.5
python-dateutil==2.9.0
